+++ web-service/src/main/java/io/acari/handler/http/api/APIPageArchiveHandler.java
@@ -12,41 +12,13 @@ import io.vertx.ext.web.RoutingContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class APIPageArchiveHandler implements Handler<RoutingContext>, Configurable<Config, APIPageArchiveHandler> {
+public class APIPageArchiveHandler extends APIBaseArchiveHandler {
   private static final Logger LOGGER = LoggerFactory.getLogger(APIPageArchiveHandler.class);
-  private final Vertx vertx;
-  private SimpleResponseHandler simpleResponseHandler;
+
 
   @Inject
   public APIPageArchiveHandler(Vertx vertx) {
-    this.vertx = vertx;
+    super(vertx, "canDelete","archive-page","name");
   }
 
-  public void handle(RoutingContext routingContext) {
-    ChainableOptional.ofNullable(routingContext.user().principal().getBoolean("canDelete", false))
-      .filter(b -> b)
-      .ifPresent(canDelete -> ChainableOptional.ofNullable(routingContext.pathParam("page"))
-        .ifPresent(name -> {
-          DeliveryOptions deliveryOptions = Config.createDeliveryOptions("archive-page");
-          JsonObject params = new JsonObject()
-            .put("name", name);
-          simpleResponseHandler.handle(routingContext, params, deliveryOptions);
-        }).orElseDo(() -> fourHundred(routingContext, "name")))
-      .orElseDo(() -> routingContext
-        .response()
-        .setStatusCode(401)
-        .end());
-  }
-
-  private void fourHundred(RoutingContext routingContext, String name) {
-    routingContext.response()
-      .setStatusCode(400)
-      .end("No " + name + " Provided, bruv.");
-  }
-
-  @Override
-  public APIPageArchiveHandler applyConfiguration(Config config) {
-    this.simpleResponseHandler = new SimpleResponseHandler(config, vertx);
-    return this;
-  }
 }
